refactor(rust): `MORA_LIST_MINIMUM`を`engine::talk`から出す
