+++ crates/voicevox_core/src/engine/talk/full_context_label.rs
@@ -5,7 +5,7 @@ use smallvec::SmallVec;
 
 use crate::AccentPhrase;
 
-use super::open_jtalk::FullcontextExtractor;
+use super::{super::mora_list::MORA_LIST_MINIMUM, open_jtalk::FullcontextExtractor};
 
 #[derive(thiserror::Error, Debug)]
 #[error("入力テキストからのフルコンテキストラベル抽出に失敗しました: {context}")]
@@ -173,7 +173,19 @@ pub fn mora_to_text(consonant: Option<&str>, vowel: &str) -> String {
         }
     );
     // もしカタカナに変換できなければ、引数で与えた文字列がそのまま返ってくる
-    super::mora2text(&mora_text).to_string()
+    mora2text(&mora_text).to_string()
+}
+
+fn mora2text(mora: &str) -> &str {
+    for &[text, consonant, vowel] in MORA_LIST_MINIMUM {
+        if mora.len() >= consonant.len()
+            && &mora[..consonant.len()] == consonant
+            && &mora[consonant.len()..] == vowel
+        {
+            return text;
+        }
+    }
+    mora
 }
 
 #[cfg(test)]
@@ -182,6 +194,7 @@ mod tests {
 
     use ::test_util::OPEN_JTALK_DIC_DIR;
     use jlabel::Label;
+    use pretty_assertions::assert_eq;
     use rstest::rstest;
     use rstest_reuse::*;
 
@@ -451,5 +464,16 @@ mod tests {
             &extract_full_context_label(&open_jtalk.0, text).unwrap(),
             accent_phrase
         );
+    }
+
+    #[rstest]
+    #[case("da", "ダ")]
+    #[case("N", "ン")]
+    #[case("cl", "ッ")]
+    #[case("sho", "ショ")]
+    #[case("u", "ウ")]
+    #[case("fail", "fail")]
+    fn test_mora2text(#[case] mora: &str, #[case] text: &str) {
+        assert_eq!(super::mora2text(mora), text);
     }
 }
